requirements.txt
Django>=3.2.20
gunicorn>=20.1.0
python-dotenv>=0.21.1
pandas>=1.3.5
requests>=2.31.0
openpyxl>=3.1.2
django-widget-tweaks>=1.4.12
psycopg2-binary>=2.9
django_rest_framework


